fix: clear cargo clippy, and the two defects it found
`never_loop` was a vestige: the Exif segment walker became a `from_fn` iterator
yielding every segment, so its inner loop could no longer iterate.

`manual_is_multiple_of` hid a panic — bit depth is not validated until after
`Plane::parse` returns, so `bits` under 8 makes `row_bytes` zero, and the
suggestion applied verbatim would have turned the divide-by-zero into a quietly
wrong answer. Guarded first. `drop(loader)` freed nothing and is gone.

+++ crates/tohdr-apple/examples/profile_decode.rs
@@ -150,7 +150,7 @@ fn half_to_f32(h: u16) -> f32 {
                 e -= 1;
             }
             let f = f & 0x3FF;
-            ((sign << 31) | (((127 - 15 + e as i32) as u32) << 23) | (f << 13)) as u32
+            (sign << 31) | (((127 - 15 + e) as u32) << 23) | (f << 13)
         }
     } else if exp == 0x1F {
         (sign << 31) | (0xFF << 23) | (frac << 13)
